Handle promise rejection in GraphResultBrowserNodeBrowser
It’s a pretty terrible way to handle the error, but at least it’s better
than nothing at all.

This could also be seen as a small step in the long journey of migrating
this code base from jQuery’s deferred style (done/fail) to native
Promise style (then) – jQuery is compatible with either, but I think the
native Promise style is much better.

Change-Id: I613826f89d99a0100519b90a679738ba1dc9e9d6