Replace grunt-jscs with grunt-eslint to remove known jscs vulnerablities
This drops another 18 vulnerabilities associated with grunt-jscs[1]
dependencies. The remaining 9 vulnerabilities are associated with
grunt-contrib-less[2] and grunt-usemin[3]. Both seem an integral part of
our build chain and both have also not been updated for years.

Maybe we should investigate an alternative to grunt.

The basis for the eslint configuration was WikibaseLexeme.

[1]: https://www.npmjs.com/package/grunt-jscs
[2]: https://www.npmjs.com/package/grunt-contrib-less
[3]: https://www.npmjs.com/package/grunt-usemin

Change-Id: I3fa44d273334ae175fdbe617a0e7f35cc1f09ef7