chore: upgrade dependencies to make npm audit happy
The changes to style.less were necessary to fix some linting errors that
appeared. They should make no difference at all.

The remaining issues are in a dependency of grunt-usemin: a old lodash
version vulnerable to Prototype Pollution. This is judged to not be an
issue for us, but the reports are annoying and grunt-usemin is entirely
unmaintained. So we should look into moving away from grunt and
modernising our build chain. T274626 was filed for that purpose.

Change-Id: Ic95ce7d8f09b9d6efaca743edea747cd8a4754df