Update title when loading query
The responsibility for updating the document.title moves from ResultView
into App, and we do it not just when running a query, but also when
loading the page (if there was a query in the URL). ResultView still
shows the action bar with the title, as before.

The code to parse the title out of the query is duplicated in both
files; it’s simple enough that I’m not convinced it’s worth extracting
somewhere. (I don’t think it would fit very well in our SparqlQuery
helper, in particular.)

A rudimentary test for _updateTitle() is included – I don’t think the
interaction with the rest of App.js (i.e. that it’s called on load and
submit) can be tested easily.

Bug: T287145
Change-Id: I5211da634011bb6ff301a9b0209a29eba0a6e602