Adjust query builder button and its popup
Order of buttons has been changed so that the query
builder button is the last one.
Popup image was removed and style adjusted, so it
is smaller and less intrusive.
Minimal width for the dropdown window set to em so it remains
relativ (as opposed to e.g. fonts, that should stay absolute
and are therefore set in px)

Bug: T296135
Change-Id: I4825364094851a5063e98a0656531f78257f6a05