Improve indentation of query hint
The SERVICE is on the first indentation level, so it should only be
indented by two spacess, not four. On the other hand, the contents of
the wikibase:around and wikibase:box service suggestions further down
are on the second indentation level (SELECT + SERVICE), so those should
stay as they are, at four spaces.

Bug: T291695
Change-Id: I4e33f61a5e58734d7d8163f7b242a314fd6e35a5