Disable SVG downloads for unsupported result views
The SVG download option only works for some result views: those that are
fully rendered as an SVG that can be downloaded as a stand-alone
document (which is mainly the chart-based result views). The following
result views don’t support a proper SVG download:

* most result views don’t use SVGs at all, e.g. Table, ImageGrid, etc.;
  in this case, the download would be a silent no-op
* TreeMap uses an SVG, but the labels are <foreignElement>s that don’t
  show up when the SVG is rendered on its own, so the downloaded SVG
  would only show colored boxes
* Map contains two SVGs (the minimap, and the tiny Flag of Ukraine in
  the attribution), but the main map is not an SVG, so the downloaded
  SVG would only show a skeleton of the minimap)

Hide the SVG download option unless one of the supported result views is
used. All the other download options (JSON, TSV, CSV, HTML table) work
for all result views, so I don’t think we need to make this a generic
“hide any download format for any result browser” mechanism; we only
need a special case for the SVG download.

With this change, a download should always work, so add a console
warning and tracking if it doesn’t (replacing the aforementioned silent
no-op); the tracking is very similar to the existing tracking for the
click that initiates the download.

This also adds a missing call to _setSelectedDisplayType() in case no
default view was specified, fixing T324841.

Bug: T311594
Bug: T324841
Change-Id: Id8ef3d83e2f5a09b9e0cc7ba38f1f316293e1ce8
Co-Authored-By: Michael Große <michael.grosse@wikimedia.de>