Refactor to use wbsearchentites for tooltips
Also, this caches the data.

In the tests, this uses a custom messages that shows the output of
complex arrays and objects as stringified JSON, because the output of
grunt-contrib-qunit is just [object Object] on the console for the
actual and expected values. This is highlighted upstream at
gruntjs/grunt-contrib-qunit#197 and a fix is being proposed at
gruntjs/grunt-contrib-qunit#198.

Change-Id: I401a8079fa759b4d3afa9c19b285d647dae51ae5