eslint: Enable indent in wikibase/
Rearrange the “when ready + config” line in a way that doesn’t require
the entire rest of the file to be indented by an additional level (as
eslint would do if we just let it fix the original code).

“Push down” the disabled indent rule into the wikibase/queryService/
subdirectory; additional changes will enable it there as well. The other
subdirectory, wikibase/tests/, is in .eslintignore, so don’t bother
disabling the indent rule there. (At some point we’ll probably want to
enable eslint there, but then we’ll want to enable the indent rule too.)

Change-Id: I0e77eda36eefccce31631a73b8844b4105422c01