Decode URIs for display
This makes especially sitelinks in non-Latin scripts vastly more
readable, for example:

- <https://my.wikipedia.org/wiki/%E1%80%9C%E1%80%80%E1%80%BA%E1%80%85%E1%80%BD%E1%80%95%E1%80%BA%E1%80%99%E1%80%BB%E1%80%AC%E1%80%B8%E1%81%8F_%E1%80%A1%E1%80%9B%E1%80%BE%E1%80%84%E1%80%BA%E1%80%9E%E1%80%81%E1%80%84%E1%80%BA>
+ <https://my.wikipedia.org/wiki/လက်စွပ်များ၏_အရှင်သခင်>

If there are any “Separator” or “Other” characters hiding in the
encoding (which I would only expect from a malicious query – I don’t
think they are possible in a legitimate sitelink title, and I wouldn’t
expect them in other URI values either), encode them again; if we can’t
determine Unicode categories (IE11), or if the URI can’t be decoded at
all, bail out and show the whole encoded URI again.

Note that, because we use decodeURI() and not decodeURIComponent(), some
characters that can appear in sitelink titles will not be decoded;
specifically, this affects the following characters:

• ? (see e.g. Q50602832)
• & (Q65117582)
• = (Q4037766)
• + (Q4544986)

But not the following characters, which the Wikibase RDF export doesn’t
encode to begin with:

• ; (Q4646735)
• / (Q4115189)
• : (Q4115189)
• @ (Q277529)
• $ (Q4540344)
• , (Q16463234)

Nor the # character, which is not possible in a MediaWiki title. In
theory, I think we could decode [=?&+] by parsing the URL and applying
decodeURIComponent() to the path component, but I don’t think that extra
complexity is necessary.

Bug: T327514
Change-Id: I3bc2829af30e8b23afe6d261e57537bafe512273