Move build scripts from Jenkins job into the repository
The scripts originate from the shell steps of the
wikidata-query-gui-build Jenkins job:
https://gerrit.wikimedia.org/g/integration/config/+/78022436b25e46bce0c56f23eebc3194369a02d8/jjb/wikidata.yaml#78

Invoking these scripts from the Jenkins job takes place in Id272e206a4f.

Note: the commit is pushed by the Jenkins job rather than the script
since we use a ssh-agent which is not accessible from within the CI
container.

Bug: T328543
Change-Id: If8869cb0f5cfda6e49cb8cfc6b1499230a9d4ee5
Co-Authored-By: Antoine Musso <hashar@free.fr>