Add mul to label service params
Change the suggested label service, and various internal queries and
tests, to add `mul` to the languages before the final fallback to `en`,
so that newly written queries will do the right thing when labels in
other languages start being replaced by `mul` labels in some languages.

Whether this is sufficient to make `mul` work in practice, or whether
we’ll have to update the label service itself to keep old queries
working, remains to be seen, I think.

Bug: T304976
Change-Id: Icbadacae78c99eaf97adcb5c1cea30551bca0818